Add files via upload
Updated version of IKO42 decoder/encoder arrangement. Channels were reordered, so they increment from top to down of the array, with front/clockwise order. It allows to select specific layers of the array for monitoring or eg. excluding/reducing gain of the layers blow floor level. The name of the file refers to top-down (TD), clockwise (CW) order of the channels.